Add files via upload
Online Quiz Management System

The Online Quiz Management System is built as a Java application using the
Eclipse IDE. It incorporates the principles of object-oriented programming (OOP).
This system enables users to create, manage and take quizzes providing a user
interface, for both quiz administrators and participants. The pdf contains the source code.